:hammer: move shared components from feature
