+++ features/shared/CategoriesTabs.tsx
@@ -1,0 +1,1 @@
+import React from 'react';
