+++ features/shared/EventsLoadingProgressSpinner.css
