+++ features/shared/EventsView.tsx
@@ -6,13 +6,7 @@ import {
   Card,
   CardContent,
   Typography,
-  /* ButtonGroup, */
 } from '@material-ui/core';
-/* import IconButton from '@material-ui/core/IconButton'; */
-/* import StarIcon from '@material-ui/icons/Star'; */
-/* import EditIcon from '@material-ui/icons/Edit'; */
-/* import AddToPhotosIcon from '@material-ui/icons/AddToPhotos'; */
-/* import DeleteIcon from '@material-ui/icons/Delete'; */
 import { AgendaEventCategory, AgendaEvent } from '../../data/dbSchema';
 import { useSelector, useDispatch } from 'react-redux';
 import { getSelectedCategory } from '../../store/selectors/getSelectedCategory';
@@ -82,30 +76,6 @@ export const EventsView: React.FC = () => {
                   <Typography gutterBottom variant="h6" component="h3">
                     {event.name}
                   </Typography>
-
-                  {/* <ButtonGroup */}
-                  {/*   color="primary" */}
-                  {/*   aria-label="outlined primary button group" */}
-                  {/* > */}
-                  {/*   <IconButton> */}
-                  {/*     <StarIcon /> */}
-                  {/*   </IconButton> */}
-
-                  {/*   <IconButton color="secondary" aria-label="add an alarm"> */}
-                  {/*     <EditIcon color="primary" /> */}
-                  {/*   </IconButton> */}
-
-                  {/*   <IconButton */}
-                  {/*     color="primary" */}
-                  {/*     aria-label="add to shopping cart" */}
-                  {/*   > */}
-                  {/*     <AddToPhotosIcon /> */}
-                  {/*   </IconButton> */}
-
-                  {/*   <IconButton aria-label="delete" color="secondary"> */}
-                  {/*     <DeleteIcon /> */}
-                  {/*   </IconButton> */}
-                  {/* </ButtonGroup> */}
                 </CardContent>
               </Card>
             );
