+++ styles/components.css
@@ -1,3 +1,3 @@
-@import '../features/view-highlighted-events/components/EventsLoadingProgressSpinner.css';
-@import '../features/view-highlighted-events/components/EventsLoadingError.css';
-@import '../features/view-highlighted-events/EventsView.css';
+@import '../features/shared/EventsLoadingProgressSpinner.css';
+@import '../features/shared/EventsLoadingError.css';
+@import '../features/shared/EventsView.css';
